Delete old jison-based parser
