+++ .changeset/sharp-colts-notice.md
@@ -1,0 +1,5 @@
+---
+"@khanacademy/kas": patch
+---
+
+Delete old jison based parser
