+++ packages/kas/src/index.js
@@ -2,3 +2,4 @@ export {libVersion} from "./version";
 
 export * from "./nodes";
 export {compare} from "./compare";
+export {parse} from "./parser";
